+++ functions.zsh
@@ -1,24 +1,12 @@
 #!/usr/bin/env zsh
 link() {
-  # options
-  while [[ "$1" =~ ^- && ! "$1" == "--" ]]; do case $1 in
-    -d1 )
-      diff=1
-      ;;
-    -f1 )
-      force=1
-      ;;
-  esac; shift; done
-
-  if [[ "$1" == '--' ]]; then shift; fi
-
   local filepath=$argv[1]
   local dotfile=${argv[2]:-.${filepath:t}}
 
   if [[ $filepath -ef ~/$dotfile ]]; then
     echo "$fg[blue][•]$reset_color $dotfile already exists"
   elif [[ -e ~/$dotfile ]]; then
-    if (( $force )); then
+    if (( $FORCE )); then
       if ln -svf "${filepath:a}" "$HOME/$dotfile"; then
         echo "$fg[green][+]$reset_color $dotfile $fg[red]force$reset_color linked"
       else
@@ -39,25 +27,13 @@ link() {
 }
 
 copy() {
-  # options
-  while [[ "$1" =~ ^- && ! "$1" == "--" ]]; do case $1 in
-    -d1 )
-      diff=1
-      ;;
-    -f1 )
-      force=1
-      ;;
-  esac; shift; done
-
-  if [[ "$1" == '--' ]]; then shift; fi
-
   local filepath=$argv[1]
   local dotfile=${argv[2]:-.${filepath:t}}
 
   if [[ -e ~/$dotfile ]]; then
     if diff -q "$filepath" "$HOME/$dotfile" &>/dev/null; then
       echo "$fg[blue][•]$reset_color $dotfile already exists"
-    elif (( $force )); then
+    elif (( $FORCE )); then
       if [[ -d $filepath ]]; then
         if cp -Rf "${filepath:a}" "$HOME/$dotfile"; then
           echo "$fg[green][+]$reset_color $dotfile directory $fg[red]force$reset_color copied"
@@ -73,7 +49,7 @@ copy() {
       fi
     else
       echo "$fg[yellow][?]$reset_color files $filepath and $dotfile differ"
-      if (( $diff )); then
+      if (( $DIFF )); then
         diff "$filepath" "$HOME/$dotfile"
       fi
     fi
@@ -95,18 +71,6 @@ copy() {
 }
 
 decipher() {
-  # options
-  while [[ "$1" =~ ^- && ! "$1" == "--" ]]; do case $1 in
-    -d1 )
-      diff=1
-      ;;
-    -f1 )
-      force=1
-      ;;
-  esac; shift; done
-
-  if [[ "$1" == '--' ]]; then shift; fi
-
   local filepath=$argv[1]
   local dotfile=${argv[2]:-.${filepath:t}}
 
@@ -115,7 +79,7 @@ decipher() {
   elif [[ -e ~/$dotfile ]]; then
     if gpg --no-tty -q -d "$filepath" | diff -q - "$HOME/$dotfile" &>/dev/null; then
       echo "$fg[blue][•]$reset_color $dotfile already exists"
-    elif (( $force )); then
+    elif (( $FORCE )); then
       if gpg --no-tty -q -d "$filepath" 1>| "$HOME/$dotfile"; then
         echo "$fg[green][+]$reset_color $dotfile file $fg[red]force$reset_color deciphered"
       else
@@ -123,7 +87,7 @@ decipher() {
       fi
     else
       echo "$fg[yellow][?]$reset_color files $filepath and $dotfile differ"
-      if (( $diff )); then
+      if (( $DIFF )); then
         gpg --no-tty -q -d "$filepath" | diff - "$HOME/$dotfile"
       fi
     fi